+++ docs/changelog.rst
@@ -5,4 +5,6 @@ Changelog
 1.0.0 (unreleased)
 ------------------
 
+- #4 Improve ASTM Protocol for Multi connections
+- #3 Fix ASTM message splitting and added tests
 - #1 Ensure output messages are LIS2-A2 compliant